==> demo/4_visualization.ipynb <==
import glob, pickle
import pathlib, os

import numpy as np
import scipy as sp
import pandas as pd

from scipy.interpolate import make_lsq_spline
import open3d as o3d
from sklearn.decomposition import PCA

import matplotlib.pyplot as plt
import seaborn as sns
import plotly.express as px
import plotly.graph_objects as go

file_path = "SPLINED_CURVE_PATH"

file_path = "data/splined_curve/demo_splined_curve.pickle"

with open(file_path, "rb") as f:
    coord_3d_recons = pickle.load(f)

%matplotlib inline
fig = plt.figure(figsize = (12, 12))
fig.patch.set_alpha(0.)
ax = fig.add_subplot(111, projection='3d')

ax.set_xlabel("x", size = 14)
ax.set_ylabel("z", size = 14)
ax.set_zlabel("y", size = 14)
ax.set_box_aspect((1,1,1))
ax.view_init(elev=30, azim=-90)

ax.axis("off")

for i in range(len(coord_3d_recons)):
    x_q, y_q, z_q = coord_3d_recons[i][:,0], coord_3d_recons[i][:,1], coord_3d_recons[i][:,2]
    ax.plot(x_q, y_q, z_q,"-",lw=3)#,color="#920092")

plt.show()
